backend/requirements.txt
hydra-core==1.3.*
numpy==2.*
omegaconf==2.3.*
onnx==1.19.*
onnxruntime==1.23.*
sentencepiece
torch>=2.5,<2.9
torchaudio>=2.5,<2.9
tqdm
huggingface_hub==1.4.1
pyannote.audio==4.0.3
python-dotenv==1.2.1
python-multipart
requests